requirements.txt
groundingdino
openai
opencv-python
torch
torchvision
Pillow
transformers
segment-anything
git+https://github.com/facebookresearch/segment-anything.git
